File iterator refactor
