+++ GLOCKFILE
@@ -1,22 +1,29 @@
-github.com/araddon/dateparse 5051e9281eea7676a49bbe2a0b3abfac8c58f59a
+cloud.google.com/go c116c7972ec94f148459a304d07a67ecbc770d4b
+github.com/araddon/dateparse 5ea40eaf943871f7af8232c1bca6b96b8af84b44
 github.com/araddon/gou 0ab1e8f997e35d27da012072dc531fcc0204da5a
 github.com/bmizerany/assert b7ed37b82869576c289d7d97fb2bbd8b64a0cb28
-github.com/dataux/dataux 0dd110826843510d88a9d6e2c14b6943a8fe5105
+github.com/dataux/dataux 3ca4d4db4990e7056ef37ac63f8a23cd6d15958c
 github.com/dchest/siphash 4ebf1de738443ea7f45f02dc394c4df1942a126d
 github.com/go-sql-driver/mysql 2e00b5cd70399450106cec6431c2e2ce3cae5034
-github.com/gogo/protobuf f9114dace7bd920b32f943b3c73fafbcbab2bf31
-github.com/golang/protobuf 8ee79997227bf9b34611aee7946ae64735e6fd93
+github.com/gogo/protobuf d66ead2957a1d163be783c18c25bc388a201cdd0
+github.com/golang/protobuf 69b215d01a5606c843240eab4937eab3acee6530
 github.com/google/btree 316fb6d3f031ae8f4d457c6c5186b9e3ded70435
-github.com/hashicorp/go-immutable-radix 76b5f4e390910df355bfb9b16b41899538594a05
-github.com/hashicorp/go-memdb d2d2b77acab85aa635614ac17ea865969f56009e
+github.com/googleapis/gax-go da06d194a00e19ce00d9011a13931c3f6f6887c7
+github.com/hashicorp/go-immutable-radix 30664b879c9a771d8d50b137ab80ee0748cb2fcc
+github.com/hashicorp/go-memdb 6b158e314030abfb0dbbbc24d0ac71b132259ebf
 github.com/hashicorp/golang-lru 0a025b7e63adc15a622f29b0b2c4c3848243bbf6
+github.com/jmoiron/sqlx f4076845477b10ac2453a16377a8383467aafe72
 github.com/kr/pretty cfb55aafdaf3ec08f0db22699ab822c50091b1c4
 github.com/kr/text 7cafcd837844e784b526369c9bce262804aebc60
 github.com/leekchan/timeutil 28917288c48df3d2c1cfe468c273e0b2adda0aa5
+github.com/lytics/cloudstorage 22e00a2922f9f61710094882323562e3ff3e64c8
 github.com/lytics/datemath 2c793e7cfff97be9feb3bfe3c249cb5a83842ebb
 github.com/lytics/logrus 59cd2e4a569b4494d9196bef82f0784e467adac9
 github.com/mb0/glob 1eb79d2de6c448664e7272f8b9fe1938239e3aaa
 github.com/pborman/uuid 1b00554d822231195d1babd97ff4a781231955c9
 github.com/surge/sqlparser 6b860f881ddbb9373d7173bdfa1f052ec3e6b215
 github.com/zhenjl/sqlparser 6b860f881ddbb9373d7173bdfa1f052ec3e6b215
-golang.org/x/net f2499483f923065a842d38eb4c7f1927e6fc6e6d
+golang.org/x/net d379faa25cbdc04d653984913a2ceb43b0bc46d7
+golang.org/x/oauth2 f6093e37b6cb4092101a298aba5d794eb570757f
+google.golang.org/api dfa61ae24628a06502b9c2805d983b57e89399b5
+google.golang.org/grpc 708a7f9f3283aa2d4f6132d287d78683babe55c8
