+++ datasource/files/README.md
@@ -1,38 +1,41 @@
 
+File Data Source
+---------------------------------
+
+Turn files into a SQL Queryable data source.
+
+Allows Cloud storage (Google Storage, S3, etc) to be queried with traditional 
+sql.  Also allows these files to have custom serializations, compressions,
+encryptions.
+
+**Design Goal**
+* *Hackable Stores* easy to add to Google Storage, local files, s3, etc.
+* *Hackable File Formats* Protbuf files, WAL files, mysql-bin-log's etc.
 
 
+**Developing new Sources**
+
+* *FileStore* defines file storage (s3, google-storage, local files, sftp, etc)
+  * *StoreReader* defines file storage reader(writer) for finding lists of files, and
+    opening files.
+* *FileHandler* Defines Registry to create `FileScanner`. Also for 
+    extracting info from path info, ie often folders serve as "columns" in the virtual table.
+  * *FileScanner* File Row Reading, how to transform contents of
+    file into *qlbridge.Message* for use in query engine.
+    Currently CSV, Json types.  
 
 
 TODO
---------------------
-* File Paging not all in memory
-  - file iterator
-    - filesource
-      - loadSchema() not loading array of files
-      - partition needs to be consistent
-* predicate push-down into api with folder path/query.  write as new generator model.
-* subscribe to updates
-* s3
-* Keep better status of pre-fetched files to see if we have "finished" them
+----------------------------
+
+* implement event-notification
+* SFTP
+* Change the store library?  Currently https://github.com/lytics/cloudstorage but consider:
+  * https://github.com/graymeta/stow
+  * https://github.com/rook/rook
+  * https://minio.io/
+  * http://rclone.org
 
 
 
 
-		if _, tableExists := m.files[fi.Table]; !tableExists {
-			u.Debugf("%p found new table: %q", m, fi.Table)
-			m.files[fi.Table] = make([]*FileInfo, 0)
-			m.tablenames = append(m.tablenames, fi.Table)
-			nextPartId = 0
-		}
-		if fi.Partition == 0 && m.ss.Conf.PartitionCt > 0 {
-			// assign a partition
-			fi.Partition = nextPartId
-			//u.Debugf("%d found file part:%d  %s", len(m.files[fi.Table]), fi.Partition, fi.Name)
-			nextPartId++
-			if nextPartId >= m.ss.Conf.PartitionCt {
-				nextPartId = 0
-			}
-		}
-
-
-
